- 隐藏女仆面板 - 修正获取多个歌曲详情时的错误问题 - 修正歌曲 id 类型，现在全部为 long 类型 - 为歌单增加跳转到外部链接的功能

+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/client/gui/item/GuiMusicList.java
@@ -8,16 +8,12 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.gui.Gui;
 import net.minecraft.client.renderer.GlStateManager;
-import net.minecraft.client.renderer.RenderHelper;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.resources.I18n;
-import net.minecraft.init.Blocks;
-import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fml.client.GuiScrollingList;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
-import org.lwjgl.opengl.GL11;
 
 import java.util.List;
 
@@ -50,7 +46,7 @@ public class GuiMusicList extends GuiScrollingList {
     @Override
     protected void elementClicked(int index, boolean doubleClick) {
         if (doubleClick) {
-            int id = playList.getTracks().get(index).getId();
+            long id = playList.getTracks().get(index).getId();
             CommonProxy.INSTANCE.sendToServer(new PortableAudioMessageToServer(parent.getAudio().getUniqueID(), id));
         } else {
             MUSIC_INDEX = index;
